requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-wtf==1.1.1
pillow==10.0.1
werkzeug==2.3.7
wtforms==3.0.1
email-validator==2.0.0
gunicorn==21.2.0
